+++ core/auth/src/main/java/software/amazon/awssdk/auth/credentials/ProfileCredentialsProvider.java
@@ -15,6 +15,7 @@
 
 package software.amazon.awssdk.auth.credentials;
 
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
@@ -23,6 +24,7 @@ import software.amazon.awssdk.annotations.SdkTestInternalApi;
 import software.amazon.awssdk.auth.credentials.internal.ProfileCredentialsUtils;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.profiles.ProfileFile;
+import software.amazon.awssdk.profiles.ProfileFileSupplier;
 import software.amazon.awssdk.profiles.ProfileFileSystemSetting;
 import software.amazon.awssdk.utils.IoUtils;
 import software.amazon.awssdk.utils.SdkAutoCloseable;
@@ -46,55 +48,41 @@ public final class ProfileCredentialsProvider
     implements AwsCredentialsProvider,
                SdkAutoCloseable,
                ToCopyableBuilder<ProfileCredentialsProvider.Builder, ProfileCredentialsProvider> {
-    private final AwsCredentialsProvider credentialsProvider;
+
+    private volatile AwsCredentialsProvider credentialsProvider;
     private final RuntimeException loadException;
-
-    private final ProfileFile profileFile;
+    private final Supplier<ProfileFile> profileFile;
+    private volatile ProfileFile currentProfileFile;
     private final String profileName;
-
     private final Supplier<ProfileFile> defaultProfileFileLoader;
 
     /**
      * @see #builder()
      */
     private ProfileCredentialsProvider(BuilderImpl builder) {
-        AwsCredentialsProvider credentialsProvider = null;
-        RuntimeException loadException = null;
-        ProfileFile profileFile = null;
-        String profileName = null;
+        this.defaultProfileFileLoader = builder.defaultProfileFileLoader;
+
+        RuntimeException thrownException = null;
+        String selectedProfileName = null;
+        Supplier<ProfileFile> selectedProfileSupplier = null;
 
         try {
-            profileName = builder.profileName != null ? builder.profileName
-                                                      : ProfileFileSystemSetting.AWS_PROFILE.getStringValueOrThrow();
+            selectedProfileName = Optional.ofNullable(builder.profileName)
+                                          .orElseGet(ProfileFileSystemSetting.AWS_PROFILE::getStringValueOrThrow);
 
-            // Load the profiles file
-            profileFile = Optional.ofNullable(builder.profileFile)
-                                  .orElseGet(builder.defaultProfileFileLoader);
+            selectedProfileSupplier = Optional.ofNullable(builder.profileFile)
+                                              .orElseGet(() -> builder.defaultProfileFileLoader);
 
-            // Load the profile and credentials provider
-            String finalProfileName = profileName;
-            ProfileFile finalProfileFile = profileFile;
-            credentialsProvider =
-                    profileFile.profile(profileName)
-                               .flatMap(p -> new ProfileCredentialsUtils(finalProfileFile, p, finalProfileFile::profile)
-                                       .credentialsProvider())
-                               .orElseThrow(() -> {
-                                   String errorMessage = String.format("Profile file contained no credentials for " +
-                                                                       "profile '%s': %s", finalProfileName, finalProfileFile);
-                                   return SdkClientException.builder().message(errorMessage).build();
-                               });
         } catch (RuntimeException e) {
             // If we couldn't load the credentials provider for some reason, save an exception describing why. This exception
-            // will only be raised on calls to getCredentials. We don't want to raise an exception here because it may be
+            // will only be raised on calls to resolveCredentials. We don't want to raise an exception here because it may be
             // expected (eg. in the default credential chain).
-            loadException = e;
+            thrownException = e;
         }
 
-        this.loadException = loadException;
-        this.credentialsProvider = credentialsProvider;
-        this.profileFile = profileFile;
-        this.profileName = profileName;
-        this.defaultProfileFileLoader = builder.defaultProfileFileLoader;
+        this.loadException = thrownException;
+        this.profileName = selectedProfileName;
+        this.profileFile = selectedProfileSupplier;
     }
 
     /**
@@ -127,14 +115,37 @@ public final class ProfileCredentialsProvider
         if (loadException != null) {
             throw loadException;
         }
+
+        ProfileFile cachedOrRefreshedProfileFile = refreshProfileFile();
+        if (isNewProfileFile(cachedOrRefreshedProfileFile)) {
+            synchronized (this) {
+                if (isNewProfileFile(cachedOrRefreshedProfileFile)) {
+                    currentProfileFile = cachedOrRefreshedProfileFile;
+                    handleProfileFileReload(cachedOrRefreshedProfileFile);
+                }
+            }
+        }
+
         return credentialsProvider.resolveCredentials();
+    }
+
+    private void handleProfileFileReload(ProfileFile profileFile) {
+        credentialsProvider = createCredentialsProvider(profileFile, profileName);
+    }
+
+    private ProfileFile refreshProfileFile() {
+        return profileFile.get();
+    }
+
+    private boolean isNewProfileFile(ProfileFile profileFile) {
+        return !Objects.equals(currentProfileFile, profileFile);
     }
 
     @Override
     public String toString() {
         return ToString.builder("ProfileCredentialsProvider")
                        .add("profileName", profileName)
-                       .add("profileFile", profileFile)
+                       .add("profileFile", currentProfileFile)
                        .build();
     }
 
@@ -150,6 +161,17 @@ public final class ProfileCredentialsProvider
         return new BuilderImpl(this);
     }
 
+    private AwsCredentialsProvider createCredentialsProvider(ProfileFile profileFile, String profileName) {
+        // Load the profile and credentials provider
+        return profileFile.profile(profileName)
+                          .flatMap(p -> new ProfileCredentialsUtils(profileFile, p, profileFile::profile).credentialsProvider())
+                          .orElseThrow(() -> {
+                              String errorMessage = String.format("Profile file contained no credentials for " +
+                                                                  "profile '%s': %s", profileName, profileFile);
+                              return SdkClientException.builder().message(errorMessage).build();
+                          });
+    }
+
     /**
      * A builder for creating a custom {@link ProfileCredentialsProvider}.
      */
@@ -158,6 +180,7 @@ public final class ProfileCredentialsProvider
         /**
          * Define the profile file that should be used by this credentials provider. By default, the
          * {@link ProfileFile#defaultProfileFile()} is used.
+         * @see #profileFile(Supplier)
          */
         Builder profileFile(ProfileFile profileFile);
 
@@ -168,6 +191,14 @@ public final class ProfileCredentialsProvider
         Builder profileFile(Consumer<ProfileFile.Builder> profileFile);
 
         /**
+         * Define the mechanism for loading profile files.
+         *
+         * @param profileFileSupplier Supplier interface for generating a ProfileFile instance.
+         * @see #profileFile(ProfileFile) 
+         */
+        Builder profileFile(Supplier<ProfileFile> profileFileSupplier);
+
+        /**
          * Define the name of the profile that should be used by this credentials provider. By default, the value in
          * {@link ProfileFileSystemSetting#AWS_PROFILE} is used.
          */
@@ -176,11 +207,12 @@ public final class ProfileCredentialsProvider
         /**
          * Create a {@link ProfileCredentialsProvider} using the configuration applied to this builder.
          */
+        @Override
         ProfileCredentialsProvider build();
     }
 
     static final class BuilderImpl implements Builder {
-        private ProfileFile profileFile;
+        private Supplier<ProfileFile> profileFile;
         private String profileName;
         private Supplier<ProfileFile> defaultProfileFileLoader = ProfileFile::defaultProfileFile;
 
@@ -188,15 +220,16 @@ public final class ProfileCredentialsProvider
         }
 
         BuilderImpl(ProfileCredentialsProvider provider) {
-            this.profileFile = provider.profileFile;
             this.profileName = provider.profileName;
             this.defaultProfileFileLoader = provider.defaultProfileFileLoader;
+            this.profileFile = provider.profileFile;
         }
 
         @Override
         public Builder profileFile(ProfileFile profileFile) {
-            this.profileFile = profileFile;
-            return this;
+            return profileFile(Optional.ofNullable(profileFile)
+                                       .map(ProfileFileSupplier::fixedProfileFile)
+                                       .orElse(null));
         }
 
         public void setProfileFile(ProfileFile profileFile) {
@@ -206,6 +239,16 @@ public final class ProfileCredentialsProvider
         @Override
         public Builder profileFile(Consumer<ProfileFile.Builder> profileFile) {
             return profileFile(ProfileFile.builder().applyMutation(profileFile).build());
+        }
+
+        @Override
+        public Builder profileFile(Supplier<ProfileFile> profileFileSupplier) {
+            this.profileFile = profileFileSupplier;
+            return this;
+        }
+
+        public void setProfileFile(Supplier<ProfileFile> supplier) {
+            profileFile(supplier);
         }
 
         @Override
@@ -225,8 +268,9 @@ public final class ProfileCredentialsProvider
 
         /**
          * Override the default configuration file to be used when the customer does not explicitly set
-         * profileName(profileName);
-         * {@link #profileFile(ProfileFile)}. Use of this method is only useful for testing the default behavior.
+         * profileFile(ProfileFile) or profileFileSupplier(supplier);
+         * {@link #profileFile(ProfileFile)}. Use of this method is
+         * only useful for testing the default behavior.
          */
         @SdkTestInternalApi
         Builder defaultProfileFileLoader(Supplier<ProfileFile> defaultProfileFileLoader) {
@@ -234,4 +278,5 @@ public final class ProfileCredentialsProvider
             return this;
         }
     }
+
 }