+++ core/auth/src/main/java/software/amazon/awssdk/auth/token/credentials/ProfileTokenProvider.java
@@ -40,53 +40,41 @@ public final class ProfileTokenProvider implements SdkTokenProvider, SdkAutoClos
     private final SdkTokenProvider tokenProvider;
     private final RuntimeException loadException;
 
-    private final ProfileFile profileFile;
     private final String profileName;
 
     /**
      * @see #builder()
      */
     private ProfileTokenProvider(BuilderImpl builder) {
-        SdkTokenProvider tokenProvider = null;
-        RuntimeException loadException = null;
-        ProfileFile profileFile = null;
-        String profileName = null;
+        SdkTokenProvider sdkTokenProvider = null;
+        RuntimeException thrownException = null;
+        Supplier<ProfileFile> selectedProfileFile = null;
+        String selectedProfileName = null;
 
         try {
-            profileName = builder.profileName != null ? builder.profileName
-                                                      : ProfileFileSystemSetting.AWS_PROFILE.getStringValueOrThrow();
+            selectedProfileName = Optional.ofNullable(builder.profileName)
+                                          .orElseGet(ProfileFileSystemSetting.AWS_PROFILE::getStringValueOrThrow);
 
             // Load the profiles file
-            profileFile = Optional.ofNullable(builder.profileFile)
-                                  .orElse(builder.defaultProfileFileLoader)
-                                  .get();
+            selectedProfileFile = Optional.ofNullable(builder.profileFile)
+                                          .orElse(builder.defaultProfileFileLoader);
 
             // Load the profile and token provider
-            String finalProfileName = profileName;
-            ProfileFile finalProfileFile = profileFile;
-            tokenProvider =
-                    profileFile.profile(profileName)
-                               .flatMap(p -> new ProfileTokenProviderLoader(finalProfileFile, p).tokenProvider())
-                               .orElseThrow(() -> {
-                                   String errorMessage = String.format("Profile file contained no information for " +
-                                                                       "profile '%s': %s", finalProfileName, finalProfileFile);
-                                   return SdkClientException.builder().message(errorMessage).build();
-                               });
+            sdkTokenProvider = createTokenProvider(selectedProfileFile, selectedProfileName);
+
         } catch (RuntimeException e) {
             // If we couldn't load the provider for some reason, save an exception describing why.
-            loadException = e;
+            thrownException = e;
         }
 
-        if (loadException != null) {
-            this.loadException = loadException;
+        if (thrownException != null) {
+            this.loadException = thrownException;
             this.tokenProvider = null;
-            this.profileFile = null;
             this.profileName = null;
         } else {
             this.loadException = null;
-            this.tokenProvider = tokenProvider;
-            this.profileFile = profileFile;
-            this.profileName = profileName;
+            this.tokenProvider = sdkTokenProvider;
+            this.profileName = selectedProfileName;
         }
     }
 
@@ -127,7 +115,6 @@ public final class ProfileTokenProvider implements SdkTokenProvider, SdkAutoClos
     public String toString() {
         return ToString.builder("ProfileTokenProvider")
                        .add("profileName", profileName)
-                       .add("profileFile", profileFile)
                        .build();
     }
 
@@ -135,6 +122,16 @@ public final class ProfileTokenProvider implements SdkTokenProvider, SdkAutoClos
     public void close() {
         // The delegate provider may be closeable. In this case, we should clean it up when this token provider is closed.
         IoUtils.closeIfCloseable(tokenProvider, null);
+    }
+
+    private SdkTokenProvider createTokenProvider(Supplier<ProfileFile> profileFile, String profileName) {
+        return new ProfileTokenProviderLoader(profileFile, profileName)
+            .tokenProvider()
+            .orElseThrow(() -> {
+                String errorMessage = String.format("Profile file contained no information for " +
+                                                    "profile '%s'", profileName);
+                return SdkClientException.builder().message(errorMessage).build();
+            });
     }
 
     /**